+++ packages/sp/index.ts
@@ -1,37 +1,19 @@
-// TODO:: scope this down once we are done refactoring
 export * from "./sharepointqueryable.js";
-
-export {
-    SPBatch,
-} from "./batch.js";
 
 export * from "./decorators.js";
 
 export * from "./operations.js";
 
 export {
-    ISPConfiguration,
-    ISPConfigurationPart,
-} from "./splibconfig.js";
-
-export {
-    SPHttpClient,
-} from "./sphttpclient.js";
-
-export {
-    SPRest,
-    sp,
-} from "./rest.js";
-
-export {
+    SPRest2,
     sp as sp2,
 } from "./rest-2.js";
 
 export * from "./types.js";
 
 export {
-    toAbsoluteUrl,
-} from "./utils/toabsoluteurl.js";
+    SPTagging,
+} from "./behaviors/telemetry-2.js";
 
 export {
     extractWebUrl,
@@ -44,6 +26,7 @@ export {
 
 export {
     odataUrlFrom,
-} from "./odata.js";
+} from "./utils/odataUrlFrom.js";
 
-export * from "./behaviors/index.js";
+export * from "./behaviors/defaults.js";
+export * from "./behaviors/telemetry-2.js";