+++ packages/sp/sharing/funcs.ts
@@ -21,7 +21,7 @@ import {
 import { spPost } from "../operations.js";
 import { tag } from "../telemetry.js";
 import { RoleDefinitions } from "../security/types.js";
-import { emptyGuid } from "../splibconfig.js";
+import { emptyGuid } from "../types.js";
 
 /**
  * Shares an object based on the supplied options