+++ packages/sp/site-users/web.ts
@@ -1,7 +1,7 @@
 import { addProp, body } from "@pnp/queryable";
 import { _Web, Web } from "../webs/types.js";
 import { ISiteUsers, SiteUsers, ISiteUser, SiteUser, IWebEnsureUserResult } from "./types.js";
-import { odataUrlFrom } from "../odata.js";
+import { odataUrlFrom } from "../utils/odataUrlFrom.js";
 import { spPost } from "../operations.js";
 
 declare module "../webs/types" {